Add files via upload
Now capable of generating both a uniform sphere of perturbed satellites, or a uniform circle of perturbed satellites in the xz plane, which then move through the gravitational field and evolve through time. Generates csv files that can be plotted in 3d using the python files. Now I need to work on the algorithm to extract the lyapunov exponents.